Refactor: 파일이름 수정 adminApi.ts >> admin-api.ts
